+++ tests/unit-global/esnext.symbol.async-dispose.js
@@ -1,11 +1,11 @@
-import { DESCRIPTORS, GLOBAL } from '../helpers/constants';
+import { DESCRIPTORS } from '../helpers/constants';
 
 QUnit.test('Symbol.asyncDispose', assert => {
   assert.true('asyncDispose' in Symbol, 'Symbol.asyncDispose available');
   assert.true(Object(Symbol.asyncDispose) instanceof Symbol, 'Symbol.asyncDispose is symbol');
   // Node 20.4.0 add `Symbol.asyncDispose`, but with incorrect descriptor
   // https://github.com/nodejs/node/issues/48699
-  if (DESCRIPTORS && GLOBAL.process?.versions?.node !== '20.4.0') {
+  if (DESCRIPTORS) {
     const descriptor = Object.getOwnPropertyDescriptor(Symbol, 'asyncDispose');
     assert.false(descriptor.enumerable, 'non-enumerable');
     assert.false(descriptor.writable, 'non-writable');