+++ tests/unit-global/esnext.symbol.dispose.js
@@ -1,11 +1,11 @@
-import { DESCRIPTORS, GLOBAL } from '../helpers/constants';
+import { DESCRIPTORS } from '../helpers/constants';
 
 QUnit.test('Symbol.dispose', assert => {
   assert.true('dispose' in Symbol, 'Symbol.dispose available');
   assert.true(Object(Symbol.dispose) instanceof Symbol, 'Symbol.dispose is symbol');
   // Node 20.4.0 add `Symbol.dispose`, but with incorrect descriptor
   // https://github.com/nodejs/node/issues/48699
-  if (DESCRIPTORS && GLOBAL.process?.versions?.node !== '20.4.0') {
+  if (DESCRIPTORS) {
     const descriptor = Object.getOwnPropertyDescriptor(Symbol, 'dispose');
     assert.false(descriptor.enumerable, 'non-enumerable');
     assert.false(descriptor.writable, 'non-writable');